refactor: ticket관련 repository 패키지 경로 리팩토링 (#144)
