+++ src/main/java/kahlua/KahluaProject/service/TicketService.java
@@ -8,7 +8,7 @@ import kahlua.KahluaProject.domain.ticket.Status;
 import kahlua.KahluaProject.domain.ticketInfo.TicketInfo;
 import kahlua.KahluaProject.dto.ticketInfo.request.TicketInfoRequest;
 import kahlua.KahluaProject.dto.ticketInfo.response.TicketInfoResponse;
-import kahlua.KahluaProject.repository.ticket.TicketInfoRepository;
+import kahlua.KahluaProject.repository.ticket.TicketInfoRepository.TicketInfoRepository;
 import kahlua.KahluaProject.vo.TicketInfoData;
 import kahlua.KahluaProject.domain.ticket.Ticket;
 import kahlua.KahluaProject.domain.ticket.Type;
@@ -18,7 +18,7 @@ import kahlua.KahluaProject.dto.ticket.request.TicketCreateRequest;
 import kahlua.KahluaProject.dto.ticket.response.*;
 import kahlua.KahluaProject.exception.GeneralException;
 import kahlua.KahluaProject.repository.ParticipantsRepository;
-import kahlua.KahluaProject.repository.ticket.TicketRepository;
+import kahlua.KahluaProject.repository.ticket.TicketRepository.TicketRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
